% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_counts.R
\name{aggregate_counts}
\alias{aggregate_counts}
\title{Aggregate counts}
\usage{
aggregate_counts(predictions_list)
}
\arguments{
\item{predictions_list}{list of predictions from model}
}
\value{
df with score threshold applied
}
\description{
aggregate counts for detections of the same class in a single image. Helper function for `deploy_model`
}
\details{
This function removes bounding box coordinates from the model output and
aggregates multiple detections of the same class per image into a classwise prediction count.
If multiple animals of the same class are detected, the lowest confidence score will be retained to 
reflect a conservative uncertainty estimate of the count.
}
