% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_score_threshold.R
\name{apply_score_threshold}
\alias{apply_score_threshold}
\title{Applies Score Threshold to Predictions and Calculates confidence that the image is empty}
\usage{
apply_score_threshold(df, file_list, score_threshold)
}
\arguments{
\item{df}{Dataframe of predictions from model}

\item{file_list}{Vector of file names processed by model}

\item{score_threshold}{Threshold score for keeping bounding boxes}
}
\value{

}
\description{
This function removes detections that fall below the score_threshold.  When an image
has all detections removed it is assigned a value of 'empty'. The confidence in the
prediction that the image is empty is calculated as 1 - max(x) where x is a vector of
confidence values for the detections that fall below score threshold. So the confidence
that the image is empty is assumed to be directly related to the largest confidence in
the detections that will be removed.
}
