% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Define image dataset}
\usage{
dataset(data_dir, recursive = TRUE, file_extensions = c(".jpg", ".JPG"))
}
\arguments{
\item{data_dir}{data directory}

\item{recursive}{boolean}

\item{file_extensions}{file extensions}
}
\value{
list of image paths for reading into model
}
\description{
Create list of image files from data_dir. Helper function for `deploy_model`
}
