% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_output.R
\name{decode_output}
\alias{decode_output}
\title{Extract relevant model output}
\usage{
decode_output(output, label_encoder)
}
\arguments{
\item{output}{this is a subset of the list output from the neural net}

\item{label_encoder}{passed from deploy model function}
}
\value{
a dataframe of model output for an image that can be interpreted in R
}
\description{
Converts the output from the neural network into a format that 
can be used in reporting results. Helper function for `deploy_model`
}
