% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_overlap.R
\name{determine_overlap}
\alias{determine_overlap}
\title{Find overlap of two bounding boxes}
\usage{
determine_overlap(a, b)
}
\arguments{
\item{a}{Bounding box that will be compared with b}

\item{b}{Second bounding box to compare with a}
}
\value{
overlap area of `a` and `b`
}
\description{
Calculates the proportion of overlap among two bounding boxes. 
Run when the user specifies an overlap correction; helper function for `find_unique_sets`
}
