% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_one_image.R
\name{eval_one_image}
\alias{eval_one_image}
\title{eval_one_image}
\usage{
eval_one_image(
  input,
  filename,
  label_encoder,
  overlap_correction,
  overlap_threshold,
  location
)
}
\arguments{
\item{label_encoder}{}

\item{location}{}
}
\value{
pred_df data frame of formatted predictions for the image
}
\description{
Evaluate a single image in the model. Helper function for `deploy_model`
}
