% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R, R/generate_sequence_data.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata helper function}
\usage{
extract_metadata(
  data_dir = NULL,
  cam_prefix = NULL,
  burst_length = 1,
  wait_time = 0
)

extract_metadata(
  data_dir = NULL,
  cam_prefix = NULL,
  burst_length = 1,
  wait_time = 0
)
}
\arguments{
\item{data_dir}{absolute path to image directory or list of absolute path of image file names. If you specify a 
directory, the function will search recursively through this directory for all jpeg files.}

\item{cam_prefix}{character vector, where each value corresponds to a unique name 
for each camera in your dataset as it appears in the absolute path. The function will use
this identifier if your camera metadata do not contain serial numbers.}

\item{burst_length}{number of images per sequence burst.}

\item{wait_time}{time between bursts in seconds. If your cameras have a 10-minute wait time 
between triggers, your wait time is 600.}

\item{files}{character vector with absolute path to image file/files}
}
\description{
extract metadata for a single image file or list of files

create df of metadata and sequence information for all images in user image directory
}
\details{
Helper function for `deploy_model`, `generate_sequences` functions

This function generates sequence information for your image dataset. You may either specify a
path to your image directory (with option to search that directory recursively) or a character vector 
listing absolute paths to all images in your dataset. 
 this will include image dimensions, timestamp, temperature, comments, camera and sequence ids.
If you provide a string vector of camera ids represented in your directory, sequence length and 
wait time between triggers, this function will also create camera and sequence identifiers based on directory
and timestamp information. This function utility depends on either the sufficient metadata fields in each image
or reasonable directory organization in order to generate accurate sequence information. Sequences with 
different burst lengths must also be run through the function separately if sequence information does
not already exist on image metadata.
}
