% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_combinations.R
\name{find_all_combinations}
\alias{find_all_combinations}
\title{Finds all potential combinations within data frame}
\usage{
find_all_combinations(df)
}
\arguments{
\item{df}{The data frame containing bounding box values and predictions}
}
\value{

}
\description{
Helper function that generates table of all possible combinations of
bounding boxes to evaluate
}
