% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sets_with_threshold.R
\name{find_sets_with_threshold}
\alias{find_sets_with_threshold}
\title{Determine polygons that overlap}
\usage{
find_sets_with_threshold(x, y, overlap_threshold)
}
\arguments{
\item{x}{sf object containing polygons}

\item{y}{sf object containing polygons to compare with x}

\item{overlap_threshold}{The threshold (proportion) used in determining which
bounding boxes are considered unique detections}
}
\value{

}
\description{
Evaluates overlapping bounding boxes.  Applies a proportional threshold and
only evaluates bounding boxes that have an overlap greater than that proportion.
}
