% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_sets.R
\name{find_unique_sets}
\alias{find_unique_sets}
\title{Find unique sets of polygons that overlap}
\usage{
find_unique_sets(df, overlap_threshold = overlap_threshold)
}
\arguments{
\item{df}{The data frame containing bounding box values and predictions}

\item{overlap_threshold}{Threshold used in determining which bounding boxes are considered unique detections}
}
\value{
Matrix with sets of polygons that overlap
}
\description{
Evaluate overlapping bounding boxes using user-specified 
threshold to determines overlap and returns sets of polygons that overlap.
Helper function for `reduce_overlapping_bboxes`
}
