% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_input.R
\name{get_model_input}
\alias{get_model_input}
\title{get_model_input}
\usage{
get_model_input(filename, w = 408, h = 307)
}
\arguments{
\item{filename}{full path to image filename}

\item{w}{image width - must correspond to model training, DO NOT CHANGE}

\item{h}{image height - must correspond to model training, DO NOT CHANGE}
}
\value{
list of two tensors image and target
}
\description{
load image and format to pass to model
}
