% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_species.R
\name{get_possible_species}
\alias{get_possible_species}
\title{Get List of Possible Species}
\usage{
get_possible_species(location, extent.data)
}
\arguments{
\item{location}{dataframe containing a single longitude and latitude value}

\item{extent.data}{auxiliary location database}
}
\description{
This function will evaluate the user provided latitude / longitude values with range
extent data for species included in the trained models and will return a list of possible species.
}
