% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples.R
\name{get_samples}
\alias{get_samples}
\title{Extract example image set}
\usage{
get_samples()
}
\description{
Load sample image set to use in examples. Folder also contains 
ground truth labels and predictions from the species_v2 model.
}
