% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_contained_in_extent.R
\name{location_contained_in_extent}
\alias{location_contained_in_extent}
\title{Evaluates if location falls within bounding box of extent}
\usage{
location_contained_in_extent(location, extent.values)
}
\arguments{
\item{location}{dataframe containing a single longitude and latitude value}

\item{extent.values}{dataframe containing xmin, xmax, ymin, ymax representing an extent (bounding box)}
}
\description{
This function will evaluate if a location (latitude / longitude) is within
a bounding box representing a geographic extent. values with range.
}
