% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\title{models}
\format{
## `models`
A data frame with 5 rows and 5 columns
\describe{
\item{\code{model_name}}{model and version name}
\item{\code{model_type}}{model group: general, family, species, or pig_only}
\item{\code{num_classes}}{number of class labels in that model (including background)}
\item{\code{description}}{string description of model (under development)}
\item{\code{URL}}{download link for model files}
}
}
\usage{
models
}
\description{
list of all models available in the package
}
\keyword{datasets}
