% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nms.R
\name{nms}
\alias{nms}
\title{Non-maximum suppression}
\usage{
nms(df, threshold, classwise = F)
}
\arguments{
\item{df}{prediction df from `decode_output`}

\item{threshold}{IoU threshold for suppressing overlapping boxes}

\item{classwise}{boolean. Apply nms across classes (FALSE) or just within class (TRUE)}
}
\value{
nms_df filtered df using non-maximum suppression
}
\description{
apply non-maximum suppression to raw predictions. 
Adapted from the torchvision nms function: https://github.com/pytorch/vision/blob/master/torchvision/ops/boxes.py
}
