% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_img_bbox.R
\name{plot_img_bbox}
\alias{plot_img_bbox}
\title{Plot image with bounding box predictions}
\usage{
plot_img_bbox(
  filename,
  plot_df,
  output_dir,
  data_dir,
  plot_label = TRUE,
  col = "red",
  lty = 1,
  lwd = 2,
  w = w,
  h = h
)
}
\arguments{
\item{filename}{The file containing the image}

\item{plot_df}{Prediction dataframe that is output from deployment}

\item{output_dir}{Desired directory to make plots}

\item{data_dir}{absolute path to images}

\item{plot_label}{boolean. Do you want the predicted category on the plot?}

\item{col}{color of the bbox (and label if `plot_label=TRUE`). See `?plot` 
for an explanation of `col`, `lwd`, and `lty`}

\item{lty}{line type of bbox}

\item{lwd}{line width of bbox}

\item{w}{The image width (in pixels) for the annotated plot.}

\item{h}{The image height (in pixels) for the annotated plot}
}
\value{
png image file in output_dir with bboxes and labels plotted
}
\description{
Create a copy of the original image with predicted bounding box 
and, optionally, the predicted category plotted on the image copy. Helper
function for `deploy_model`
}
