% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{remove_na}
\alias{remove_na}
\title{Remove NA columns}
\usage{
remove_na(df)
}
\arguments{
\item{df}{df to filter}
}
\value{
df with all NA columns removed
}
\description{
Remove NA columns
}
\details{
Helper function to remove columns with all NA values
}
