% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{A wrapper function to run Shiny Apps from \code{CameraTrapDetectoR}.}
\usage{
runShiny(app = "deploy")
}
\arguments{
\item{app}{The name of the app you want to run. The options are currently 
`deploy`.}
}
\value{

}
\description{
Running this function will launch a shiny application to deploy object detection models
 on camera trap images
}
