% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_relabel.R
\name{smart_relabel}
\alias{smart_relabel}
\title{Relabel predictions using location}
\usage{
smart_relabel(pred_df, possible.labels, label_encoder)
}
\arguments{
\item{pred_df}{dataframe containing predictions for image}

\item{possible.labels}{dataframe containing possible labels based on species occurrence}

\item{label_encoder}{label dictionary for model being run}
}
\value{
relabeled predictions based on location
}
\description{
This function compares the labels provided by the model with a list of possible
species based on species ranges.  In some cases it relabels mis-classifications
resulting from assumption that all species could be present in photo. Helper
function for `deploy_model`
}
