% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_images.R
\name{sort_images}
\alias{sort_images}
\title{Organize image files into new directory structure sortd by class}
\usage{
sort_images(
  results = NULL,
  output_dir = NULL,
  review_threshold = NA,
  count_subdir = FALSE,
  remove_originals = FALSE
)
}
\arguments{
\item{results}{df of output from deploy_model. Currently only supports 'long' prediction format}

\item{output_dir}{absolute path to directory in which to put sorted image folders}

\item{review_threshold}{images with detections below this threshold will be moved into a folder
titled "Manual_Review". Accepts values 0.01 - 1; default = NA.}

\item{count_subdir}{boolean. Create subfolders inside each class directory for individual counts.}

\item{remove_originals}{boolean. Delete original image files. Default = FALSE}
}
\description{
make copies of image files into a new directory with a folder for each output class
}
\details{
this function takes your output results from deploy_model and sorts images into a new 
directory based on output class, with a separate folder for each class. If multiple classes are 
detected in a single image, copies of that image will be placed in every folder corresponding 
to a predicted detection. Also returns a df of the deploy_model output with a new column for 
updated filepath. Optionally removes original images. Optionally puts all images with detections 
below a certain threshold into a separate folder for manual review.
}
