% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_extent_data.R
\docType{data}
\name{species.extent.data}
\alias{species.extent.data}
\title{Documentation for species extent data}
\description{
Documentation for species extent data
}
\references{
internal
}
\author{
Ryan S Miller
}
\keyword{data}
