% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_extent_loader.R
\name{species_extent_loader}
\alias{species_extent_loader}
\title{Load data describing geographic extent of species presence}
\usage{
species_extent_loader()
}
\description{
Load data describing geographic extent of species presence
}
