% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_loader.R
\name{weight_loader}
\alias{weight_loader}
\title{Initiate model}
\usage{
weight_loader(folder)
}
\arguments{
\item{folder}{path to folder containing model files}
}
\value{
loaded model
}
\description{
Initiate model architecture and load model weights. Helper
function for `deploy_model`
}
