% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightLoader.R
\name{weightLoader}
\alias{weightLoader}
\title{Load model weights}
\usage{
weightLoader(model_type = "general", num_classes = 5)
}
\arguments{
\item{model_type}{The type of model: general, family, species, or pig-only.}

\item{num_classes}{The number of classes in the model}
}
\value{
model weights
}
\description{
Load model weights
}
