% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_output.R
\name{wide_output}
\alias{wide_output}
\title{-UNDER DEVELOPMENT- Convert output to wide format}
\usage{
wide_output(df)
}
\arguments{
\item{df}{df of predictions returned from `deploy_model` function}
}
\value{
df with output widened to facilitate counts
}
\description{
Convert output from `deploy_model` to wide format
}
