% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bbox_df.R
\name{write_bbox_df}
\alias{write_bbox_df}
\title{separate function to write bboxes to df}
\usage{
write_bbox_df(full_df, w, h, bboxes, score_threshold)
}
\arguments{
\item{full_df}{prediction output from apply_score_threshold}

\item{bboxes}{existing predicted boxes df}

\item{score_threshold}{confidence score to filter bounding boxes}
}
\description{
helper function for `deploy_model` to create bbox df for saving in checkpoint, csv
}
