% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_exif_tags.R
\name{write_exif_tags}
\alias{write_exif_tags}
\title{Write results to image metadata}
\usage{
write_exif_tags(class, conf, count, review_threshold)
}
\arguments{
\item{class}{predicted class}

\item{conf}{prediction confidence score}

\item{count}{prediction count}

\item{review_threshold}{review threshold for accepting predictions}
}
\value{
call for exifr::exiftool_call
}
\description{
Write results to image metadata
}
\details{
write exiftool config file
}
