% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_metadata_tags.R
\name{write_metadata_tags}
\alias{write_metadata_tags}
\title{Write predictions to image metadata}
\usage{
write_metadata_tags(pred_df, model_version, review_threshold)
}
\arguments{
\item{pred_df}{results for a given image}

\item{model_version}{model version so we know which tags to write}

\item{review_threshold}{review threshold for accepting predictions}
}
\description{
-UNDER DEVELOPMENT- write model outputs to image metadata
}
