% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoader.R
\name{dataLoader}
\alias{dataLoader}
\title{This function loads a batch of data to be passed to the model}
\usage{
dataLoader(file_list, index, w = 408, h = 307)
}
\arguments{
\item{file_list}{passed from deploy function}

\item{index}{i value from deploy function}

\item{w}{image width after resizing. Recommend not changing this}

\item{h}{image height after resizing. Recommend not changing this}
}
\value{
loads data from user input and transforms it into tensors to pass to the model
}
\description{
This function loads a batch of data to be passed to the model
}
