% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Create the dataset}
\usage{
dataset(
  data_dir,
  recursive = TRUE,
  file_extensions = c(".jpg", ".JPG"),
  labeled = FALSE
)
}
\arguments{
\item{data_dir}{data directory}

\item{recursive}{boolean}

\item{file_extensions}{file extensions}

\item{labeled}{boolean. not functional yet}
}
\value{

}
\description{
This function sets up the dataset
}
