% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_ouput.R
\name{decode_output}
\alias{decode_output}
\title{Decode the output from neural network}
\usage{
decode_output(output, label_encoder, h, score_threshold = 0.7)
}
\arguments{
\item{output}{this is a subset of the list output from the neural net}

\item{label_encoder}{passed from deploy model function}

\item{h}{image height after resizing. Recommend not changing this}

\item{score_threshold}{threshold score for keeping bounding boxes}
}
\value{

}
\description{
This function will convert the output from NN into a format that can be used
to make plots and provide a csv with information
}
