% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_model.R
\name{deploy_model}
\alias{deploy_model}
\title{Deploy model on camera trap images}
\usage{
deploy_model(
  data_dir = NULL,
  model_type = "general",
  recursive = TRUE,
  file_extensions = c(".jpg", ".JPG"),
  make_plots = TRUE,
  plot_label = TRUE,
  output_dir = NULL,
  sample50 = FALSE,
  write_bbox_csv = FALSE,
  score_threshold = 0.6,
  h = 307,
  w = 408,
  lty = 1,
  lwd = 2,
  col = "red",
  labeled = FALSE
)
}
\arguments{
\item{data_dir}{Absolute path to the folder containing your images}

\item{model_type}{Options are c('general', 'species', 'family'). The `general`
model predicts to the level of mammal, bird, humans, vehicles. The `species` 
model recognizes 58 species. The `family` model recognizes 25 families.}

\item{recursive}{boolean. Do you have images in subfolders within your
data_dir that you want to analyze, if so, set to TRUE. If you only want to 
analyze images within your data_dir and not within sub-folders, set to FALSE.}

\item{file_extensions}{The types of extensions on your image files. Default 
is c(".jpg", ".JPG")}

\item{make_plots}{boolean. Do you want to make plots of the images with
their predicted bounding boxes?}

\item{plot_label}{boolean. Do you want the plots to contain the predicted
class of object}

\item{output_dir}{You can specify absolute path to output. Default is `NULL`,
and creates a folder within your data_dir. Only specify a path if you want the
results stored somewhere else on your computer.}

\item{sample50}{boolean. Do you want to run the model only on a subset of 
50 images from your dataset? This is a good idea if you are experimenting 
with settings.}

\item{write_bbox_csv}{boolean. Do you want to create a csv with all of the 
information on predicted bounding boxes? This csv will include all bounding boxes,
even those with low probability.}

\item{score_threshold}{Confidence threshold for using a bounding box. 
Default is 0.6. A lower number will produce more bboxes (it will be less
stringent in deciding to make a bbox). A higher number will produce fewer
bboxes (it will be more stringent).}

\item{h}{The image height (in pixels) for the annotated plot. Only used if
\code{make_plots=TRUE}.}

\item{w}{The image width (in pixels) for the annotated plot.}

\item{lty}{line type for bbox plot. See \code{?plot} for details}

\item{lwd}{line width for bbox plot. See \code{?plot} for details}

\item{col}{line color for bbox plot. See \code{?plot} for details}

\item{labeled}{This is not functional}
}
\value{
Returns a dataframe of predictions for each file. The rows in this 
 dataframe are the file names in your `data_dir`; the columns are the categories
 in the model. If any of your images were not loaded properly, there will be a 
 column in the dataframe called `image_error`. Images with a 1 in this column 
 had issues and the model was not deployed on them.
}
\description{
This function deploys a model trained to identify and count the objects
in camera trap images.
}
\details{
This function deploys a model to detect and classify objects in camera 
 trap images. The function will find all files matching the `file_extension`s 
 specified within the `data_dir` specified and deploy the `model_type` on these
 images. If you specify \code{recusive=TRUE}, the function will find relevant image
 files within all subdirectories of your `data_dir`. `deploy_model` returns a
 dataframe of predicted number of individuals within each category in each
 image. This dataframe is also written as a csv file within your `output_dir`.
 If you specify \code{make_plots=TRUE}, the function will plot predicted bounding
 boxes for each image in your `output_dir`. If you are working with many images, 
 you may wish to specify \code{sample50=TRUE} the first time you use this function,
 which will only deploy the model on 50 of your images. There are three options for
 \code{model_type}: 'general' recognizes mammals, birds, humans, and vehicles. 
 'species' recognizes 58 species. 'family' recognizes 25 families. If you want
 to see all of the information for each bounding box (including coordinates, 
 labels, and confidence), specify \code{write_bbox_csv=TRUE} and it will be
 produced in your `output_dir`. Additionally,
 A file called "arguments" will be produced in your `output_dir` this is a list
 of all of the arguments you passed to this function for reference.
}
