% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_overlap.R
\name{determine_overlap}
\alias{determine_overlap}
\title{Finds overlap of two bounding boxes}
\usage{
determine_overlap(a, b)
}
\arguments{
\item{a}{Bounding box that will be compared with b}

\item{b}{Bounding box}
}
\value{

}
\description{
Helper function that returns the proportion of overlap among two bounding boxes
}
