% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cache.R
\name{download_cache}
\alias{download_cache}
\title{Download model files}
\usage{
download_cache(model_type = "species")
}
\arguments{
\item{model_type}{}
}
\value{
path to folder containing model files
}
\description{
Download files necessary to run the model and store them in
the package cache. Helper function for `deploy_model`
}
