% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_models.R
\name{download_models}
\alias{download_models}
\title{Download model files.}
\usage{
download_models(models)
}
\arguments{
\item{models}{user-specified model or list of models. Defaults to 
latest version of model type if no version given}
}
\value{
path to folder containing model files
}
\description{
Download files necessary to run the model and store them in
the package cache. May be run by the user to download models prior to running
`deploy_model`. Accepts a character string or vector of character strings 
associated with model types.
}
