% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_one_image.R
\name{eval_one_image}
\alias{eval_one_image}
\title{eval_one_image}
\usage{
eval_one_image(
  input,
  filename,
  label_encoder,
  score_threshold,
  overlap_correction,
  overlap_threshold,
  location,
  possible_labels,
  model,
  model_version
)
}
\arguments{
\item{input}{torch tensor. formatted image to pass to the model}

\item{filename}{string. full path to image file}

\item{label_encoder}{dict. class label dictionary}

\item{score_threshold}{numeric. arg passed from `deploy_model` to filter low confidence predictions}

\item{overlap_correction}{boolean. arg passed from `deploy_model` to apply non-maximum suppression}

\item{overlap_threshold}{numeric. arg passed from `deploy_model`, IoU threshold for NMS}

\item{location}{lat-lon input from users}

\item{possible_labels}{output of possible classes present at given location}
}
\value{
pred_df data frame of formatted predictions for the image
}
\description{
Evaluate a single image in the model. Helper function for `deploy_model`
}
