% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract sequence and metadata info and assign info to image filename}
\usage{
extract_metadata(
  data_dir = NULL,
  file_extensions = ".jpg",
  recursive = TRUE,
  burst_length = 1,
  wait_time = 0,
  cam_prefix = NULL
)
}
\arguments{
\item{data_dir}{absolute path to image directory}

\item{file_extensions}{The types of extensions on your image files. Case insensitive; enter as a string.
Accepts the following file types: ".jpg", ".png", ".tif", ".pdf". Default is ".jpg"}

\item{recursive}{boolean. Do you have images in subfolders within your
data_dir that you want to analyze, if so, set to TRUE. If you only want to 
analyze images within your data_dir and not within sub-folders, set to FALSE. Default is TRUE.}

\item{burst_length}{number of images per sequence burst.}

\item{wait_time}{time between bursts in seconds. If your cameras have a 10-minute wait time 
between triggers, your wait time is 600.}

\item{cam_prefix}{character vector, where each value corresponds to a unique name 
for each camera in your dataset as it appears in the absolute path. The function will use
this identifier if your camera metadata do not contain serial numbers.}
}
\description{
create df of metadata and sequence information for all images in user image directory
}
\details{
This function extracts and stores image-level metadata for your dataset. Depending on your
camera metadata, this will include image dimensions, timestamp, temperature, comments, camera and sequence ids.
If you provide a string vector of camera ids represented in your directory, sequence length and 
wait time between triggers, this function will also create camera and sequence identifiers based on directory
and timestamp information. This function utility depends on either the sufficient metadata fields in each image
or reasonable directory organization in order to generate accurate sequence information.
}
