% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata helper function}
\usage{
extract_metadata(files)
}
\arguments{
\item{files}{character vector with absolute path to image file/files}
}
\description{
extract metadata for a single image file or list of files
}
\details{
Helper function for `deploy_model`, `generate_sequences` functions
}
