% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_input.R
\name{get_model_input}
\alias{get_model_input}
\title{get_model_input}
\usage{
get_model_input(filename, model_version)
}
\arguments{
\item{filename}{full path to image filename}

\item{model_version}{model version; input differs for V3}
}
\value{
list of two tensors image and target
}
\description{
load image and format to pass to model
}
