% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_species.R
\name{get_possible_species}
\alias{get_possible_species}
\title{Get List of Possible Species}
\usage{
get_possible_species(location, extent.data, model_type)
}
\arguments{
\item{location}{dataframe containing a single longitude and latitude value}

\item{extent.data}{auxiliary location database}

\item{model_type}{model type or version to pull class dict labels}
}
\value{
possible class label list
}
\description{
Evaluate user-provided latitude / longitude values with range
extent data for species included in the trained models. Helper function for
`deploy_model`
}
