% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_possible_species.R
\name{location_contained_in_extent}
\alias{location_contained_in_extent}
\title{Determine if location falls within a species range}
\usage{
location_contained_in_extent(location, extent.values)
}
\arguments{
\item{location}{dataframe containing a single longitude and latitude value}

\item{extent.values}{dataframe containing xmin, xmax, ymin, ymax representing an extent (bounding box)}
}
\value{
location extent helper function
}
\description{
Evaluate if a location (latitude / longitude) is within a geographic 
extent. Helper function for `get_possible_species`
}
