% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_img_bbox.R
\name{plot_img_bbox}
\alias{plot_img_bbox}
\title{Plot image with bounding box predictions}
\usage{
plot_img_bbox(filename, plot_df, arg_list)
}
\arguments{
\item{filename}{The file containing the image}

\item{plot_df}{Prediction dataframe that is output from deployment}

\item{arg_list}{argument list to pull all the other arguments}
}
\value{
png image file in output_dir with bboxes and labels plotted
}
\description{
Create a copy of the original image with predicted bounding box 
and, optionally, the predicted category plotted on the image copy. Helper
function for `deploy_model`
}
