% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_img_bbox.R
\name{plot_img_bbox}
\alias{plot_img_bbox}
\title{Make plots of the image with bounnding box predictions}
\usage{
plot_img_bbox(
  filename,
  pred_df,
  output_dir,
  data_dir,
  plot_label = TRUE,
  col = "red",
  lty = 1,
  lwd = 2,
  prop_bbox = FALSE,
  w = 408,
  h = 307
)
}
\arguments{
\item{filename}{The file containing the image}

\item{pred_df}{Prediction dataframe that is output from deployment}

\item{output_dir}{Desired directory to make plots}

\item{data_dir}{absolute path to images}

\item{plot_label}{boolean. Do you want the predicted category on the plot?}

\item{col}{color of the bbox (and label if `plot_label=TRUE`). See `?plot` 
for an explanation of `col`, `lwd`, and `lty`}

\item{lty}{line type of bbox}

\item{lwd}{line width of bbox}

\item{prop_bbox}{boolean. Are the bbox coordinates in proportion instead of 
exact coordinates? Only `TRUE` if you are using a different image size}

\item{w}{The image width (in pixels) for the annotated plot.}

\item{h}{The image height (in pixels) for the annotated plot. Only used if
\code{make_plots=TRUE}.}
}
\value{

}
\description{
Plots original image with predicted bounding box and (optionally)
the predicted category
}
