% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_checkpoint.R
\name{save_checkpoint}
\alias{save_checkpoint}
\title{Save model output every ten images to retain results if something happens to R session mid-run}
\usage{
save_checkpoint(out_df, output_dir, write_bbox_csv)
}
\arguments{
\item{out_df}{formatted df to write to csv}

\item{output_dir}{directory to save output}

\item{write_bbox_csv}{bool indicator to create df of all bbox predictions}
}
\description{
Save model output every ten images to retain results if something happens to R session mid-run
}
