% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species.extent.data}
\alias{species.extent.data}
\title{CameraTrapDetectoR  data}
\format{
##`species.extent.data`
a data frame with 120397 rows and 7 columns
\describe{
 \item{taxa}{taxonomic level}
 \item{label}{label name}
 \item{model_type}{model type}
 \item{xmin}{xmin of polygon}
 \item{ymin}{ymin of polygon}
 \item{xmax}{xmax of polygon}
 \item{ymax}{ymax of polygon}
}
}
\usage{
species.extent.data
}
\description{
species.extent.data
}
\details{
Location polygons detailing range data for each species across taxonomic level in the models
}
\keyword{datasets}
