% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightLoader.R
\name{weightLoader}
\alias{weightLoader}
\title{Load the trained model}
\usage{
weightLoader(model_type = "general", num_classes = 5)
}
\arguments{
\item{model_type}{The type of model you want to deploy:
c('mammalBirdVehicle', '')}

\item{num_classes}{The number of classes in the model}
}
\value{

}
\description{
Loads the weights for the pretrained model.
}
