% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_output.R
\name{wide_output}
\alias{wide_output}
\title{Convert output to wide format}
\usage{
wide_output(df)
}
\arguments{
\item{df}{df of predictions returned from `deploy_model` function}
}
\description{
Convert output to wide format
}
\details{
Under development! Function to convert model output to wide format.
}
