% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bbox_df.R
\name{write_bbox_df}
\alias{write_bbox_df}
\title{separate function to write bboxes to df}
\usage{
write_bbox_df(predictions_list, w, h, bboxes, score_threshold)
}
\arguments{
\item{predictions_list}{prediction output}

\item{w}{image width}

\item{h}{image height}

\item{bboxes}{existing predicted boxes df}

\item{score_threshold}{score threshold to filter bboxes}
}
\description{
separate function to write bboxes to df
}
\details{
create bbox df for saving in checkpoint, csv
}
