% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bbox_df.R
\name{write_bbox_df}
\alias{write_bbox_df}
\title{separate function to write bboxes to df}
\usage{
write_bbox_df(predictions_list, bboxes)
}
\arguments{
\item{predictions_list}{list of current image predictions}

\item{bboxes}{existing predicted boxes df}
}
\description{
helper function for `deploy_model` to create bbox df for saving in checkpoint, csv
}
