% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_output.R, R/write_output.R
\name{write_output}
\alias{write_output}
\title{Convert output to wide format}
\usage{
write_output(full_df)

write_output(full_df)
}
\arguments{
\item{full_df}{predictions list filtered by score_threshold}

\item{df}{df of predictions returned from `deploy_model` function}
}
\description{
Convert output to wide format

Write model output
}
\details{
Under development! Function to convert model output to wide format.

Finalize model predictions in df and csv output
}
