% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write model output}
\usage{
write_output(full_df, prediction_format, label_encoder)
}
\arguments{
\item{full_df}{df filtered by score_threshold}

\item{prediction_format}{wide or long}

\item{label_encoder}{class label dict}
}
\description{
Write model output
}
\details{
Finalize model predictions in df and csv output
}
