% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_output.R
\name{write_output}
\alias{write_output}
\title{Write model output}
\usage{
write_output(full_df)
}
\arguments{
\item{full_df}{predictions list filtered by score_threshold}
}
\description{
Write model output
}
\details{
Finalize model predictions in df and csv output
}
